Restructure the files for unpaid infringements
